chore: extract AccessLevelDropdown to its own component
